+++ src/vu1_monitor/config/__init__.py
@@ -1,0 +1,3 @@
+from vu1_monitor.config.settings import settings
+
+__all__ = ["settings"]
