+++ src/vu1_monitor/config/settings.py
@@ -11,6 +11,7 @@ settings = Dynaconf(
     environments=True,
     env_switcher="VU1_ENV",
     validators=[
+        Validator("name", default="VU1-Monitor"),
         # server
         Validator("server.hostname", default="localhost"),
         Validator("server.port", default=5340),
