+++ src/vu1_monitor/dials/__init__.py
@@ -1,0 +1,3 @@
+from vu1_monitor.dials.client import VU1Client
+
+__all__ = ["VU1Client"]
