+++ src/vu1_monitor/exceptions/__init__.py
@@ -1,0 +1,7 @@
+from vu1_monitor.exceptions.dials import (
+    DialNotFound,
+    DialNotImplemented,
+    ServerNotFound,
+)
+
+__all__ = ["DialNotFound", "DialNotImplemented", "ServerNotFound"]
