+++ src/vu1_monitor/logger/__init__.py
@@ -1,0 +1,3 @@
+from vu1_monitor.logger.logger import create_logger
+
+__all__ = ["create_logger"]
