+++ src/vu1_monitor/metrics/__init__.py
@@ -1,0 +1,1 @@
+from vu1_monitor.metrics.gpu import get_gpu_utilisation
