+++ src/vu1_monitor/models/__init__.py
@@ -1,0 +1,19 @@
+from vu1_monitor.models.models import (
+    Bright,
+    Colours,
+    Dial,
+    DialImage,
+    DialType,
+    Element,
+    GPUBackend,
+)
+
+__all__ = [
+    "Bright",
+    "Colours",
+    "Dial",
+    "DialImage",
+    "DialType",
+    "Element",
+    "GPUBackend",
+]
